Extract shell exec environment helper
